Split ui into components and pages
