+++ collection.html
@@ -20,6 +20,6 @@
 	<body style="overflow: auto">
 		<div style="margin-bottom: 3rem" id="Cards"></div>
 
-		<script async src="/src/ui/collection.js" type="module"></script>
+		<script async src="/src/ui/components/collection.js" type="module"></script>
 	</body>
 </html>
