+++ src/ui/pages/start-room.js
@@ -1,4 +1,4 @@
-import {html, Component} from './lib.js'
+import {html, Component} from '../lib.js'
 
 export default class StartRoom extends Component {
 	render() {
