+++ stats.html
@@ -20,6 +20,6 @@
 	<body style="overflow: auto">
 		<div id="root"></div>
 
-		<script async src="/src/ui/stats.js" type="module"></script>
+		<script async src="/src/ui/pages/stats.js" type="module"></script>
 	</body>
 </html>
